split the runtime to different modules
