+++ ic-kit-runtime/Cargo.toml
@@ -14,6 +14,8 @@ include = ["src", "Cargo.toml", "README.md"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
 ic-kit-sys = {path="../ic-kit-sys", version="0.1.1", features=["runtime"]}
+ic-types = "0.4.1"
 tokio = {version="1.20", features=["sync", "macros"]}
+memmap = "0.7.0"
 futures = "0.3"
 actix = "0.13"
